WIP - partway through debugging pool synchronization issue
At this point, I realize that we need to limit access to runtime objects
via proxies. Otherwise, different versions of the objects (even inside Atoms)
get persisted in different runs (which are frozen independently).

Another problem proxies will address is circular references. When, for example,
a run contains a reference to itself, though the freezing mechanism only stores
an id, upon thawing, an attempt is made to retrieve the real object from storage.
But since, we're in the process of thawing the run, we end up with an infinite recursion.

The way around both of these problems is to persist a proxy object which only stores
the id of the Run or Pool. That way, recursive loads can be detected and avoided by
simply returning the proxy.